# Implement queue locallly