This PR removes the external dep and implements the queue locally, reducing allocs per op by ~50% and increasing throughput by >500%.

## Before
```
❯ go test -bench=BenchmarkQueue ./storage/internal/queue_test.go --benchmem --benchtime=
10s
goos: linux
goarch: amd64
cpu: AMD Ryzen Threadripper PRO 3975WX 32-Cores
BenchmarkQueue-30           3322           3446318 ns/op         1538364 B/op      31809 allocs/op
PASS
```

## After
```
❯ go test -bench=BenchmarkQueue ./storage/internal/queue_test.go --benchmem --benchtime=10s
goos: linux
goarch: amd64
cpu: AMD Ryzen Threadripper PRO 3975WX 32-Cores
BenchmarkQueue-30          17703            646184 ns/op          502921 B/op      14372 allocs/op
PASS
```